Update files based on review comments
