+++ invokeai/backend/inferencePipeline.py
@@ -1,12 +1,12 @@
 """
-This file implements base class Inference Model for model selection
+This file implements base class for Inference pipeline
 Implements abstract methods for inference related operations
 """
 from abc import ABC, abstractmethod
 class inferencePipeline(ABC):
 
     """
-    Instantiation of Inference model class
+    Instantiation of Inference Pipeline class
     """
 
     @abstractmethod
